refactor array problems - rename package practice to curated
