+++ src/strivers/arrays/curated/MaxProductSubArray.java
@@ -1,4 +1,4 @@
-package strivers.arrays.practice;
+package strivers.arrays.curated;
 
 import java.util.Scanner;
